Rename LinearBasicDisplay->LinearBareDisplay and LinearFeatureDisplay->LinearBasicDisplay
