+++ plugins/linear-genome-view/src/LinearBareDisplay/configSchema.ts
@@ -1,0 +1,14 @@
+import PluginManager from '@jbrowse/core/PluginManager'
+import { ConfigurationSchema } from '@jbrowse/core/configuration'
+import { baseLinearDisplayConfigSchema } from '../BaseLinearDisplay'
+
+export function configSchemaFactory(pluginManager: PluginManager) {
+  return ConfigurationSchema(
+    'LinearBareDisplay',
+    { renderer: pluginManager.pluggableConfigSchemaType('renderer') },
+    {
+      baseConfiguration: baseLinearDisplayConfigSchema,
+      explicitlyTyped: true,
+    },
+  )
+}
