+++ plugins/linear-genome-view/src/LinearBareDisplay/index.test.js
