+++ plugins/linear-genome-view/src/LinearBareDisplay/index.ts
@@ -1,0 +1,2 @@
+export { configSchemaFactory } from './configSchema'
+export { stateModelFactory } from './model'
