+++ plugins/linear-genome-view/src/LinearBasicDisplay/components/SetMaxHeight.tsx
