+++ plugins/linear-genome-view/src/LinearBasicDisplay/configSchema.ts
@@ -1,14 +1,19 @@
-import PluginManager from '@jbrowse/core/PluginManager'
 import { ConfigurationSchema } from '@jbrowse/core/configuration'
+import PluginManager from '@jbrowse/core/PluginManager'
 import { baseLinearDisplayConfigSchema } from '../BaseLinearDisplay'
 
-export function configSchemaFactory(pluginManager: PluginManager) {
+export default function configSchemaFactory(pluginManager: PluginManager) {
   return ConfigurationSchema(
     'LinearBasicDisplay',
-    { renderer: pluginManager.pluggableConfigSchemaType('renderer') },
     {
-      baseConfiguration: baseLinearDisplayConfigSchema,
-      explicitlyTyped: true,
+      renderer: pluginManager.pluggableConfigSchemaType('renderer'),
+      // overrides base
+      maxDisplayedBpPerPx: {
+        type: 'number',
+        description: 'maximum bpPerPx that is displayed in the view',
+        defaultValue: 1000,
+      },
     },
+    { baseConfiguration: baseLinearDisplayConfigSchema, explicitlyTyped: true },
   )
 }
