--- plugins/linear-genome-view/src/LinearFeatureDisplay/index.ts
@@ -1,2 +1,0 @@
-export { default as modelFactory } from './model'
-export { default as configSchema } from './configSchema'
